+++ doc/changelog.rst
@@ -4,7 +4,9 @@ Changelog
 Version 0.3 (development version)
 ---------------------------------
 
-* :func:`wrap_greenthread` now also works on greenlet objects.
+* :func:`wrap_greenthread` now raises an exception if the greenthread is
+  running or already finished. In debug mode, the exception is not more logged
+  to sys.stderr for greenthreads.
 
 2014-10-21: version 0.2
 -----------------------