+++ doc/changelog.rst
@@ -7,6 +7,7 @@ Version 0.3 (development version)
 * :func:`wrap_greenthread` now raises an exception if the greenthread is
   running or already finished. In debug mode, the exception is not more logged
   to sys.stderr for greenthreads.
+* :func:`link_future` now accepts coroutine objects.
 * :func:`link_future` now raises an exception if it is called from the
   greenthread of the aiogreen event loop.
 * Fix eventlet detection of blocking tasks: cancel the alarm when the aiogreen